완성 weekly quiz01 (use BigDecimal)

+++ src/weekly_quiz_01/Beauty.java
@@ -1,7 +1,9 @@
 package weekly_quiz_01;
 
+import java.math.BigDecimal;
+
 public class Beauty extends Product {
-    public Beauty(String name, int price, int weight) {
+    public Beauty(String name, BigDecimal price, BigDecimal weight) {
         super(name, price, weight);
     }
 }
