+++ src/weekly_quiz_01/Promotion.java
@@ -1,5 +1,7 @@
 package weekly_quiz_01;
 
+import java.math.BigDecimal;
+
 public interface Promotion {
-    public int getDiscountAmount();
+    public BigDecimal getDiscountAmount();
 }
